move process user info to demo shell
