+++ demo-shell/src/app/components/app-layout/user-info/process-user-info/process-user-info.component.scss
